split up in 3 files and reduced code complexity of getDatesForWeek
